requirements.txt
arango-rdf>=1.2.0
python-arango>=8.2.0
rdflib>=7.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
jupyter>=1.0.0 